third_party/blink/web_tests/virtual/do-not-hide-visited-color/external/wpt/css/css-color/color-mix-currentcolor-visited-getcomputedstyle-expected.txt

This is a testharness.js-based test.
[FAIL] Property background-color value 'color-mix(in srgb, currentcolor, white)' should not leak :visited for computed style
  assert_equals: expected "color(srgb 1 0.5 0.5)" but got "color(srgb 0.5 0.75098 0.5)"
Harness: the test ran to completion.
